Completa algunos servicios y entidades

+++ src/main/java/com/iait/interfaces/Linea.java
@@ -14,9 +14,9 @@ public interface Linea {
     
     BigDecimal getTasaMaxima();
     
-    Long getCuotasMinimas();
+    Integer getCuotasMinimas();
     
-    Long getCuotasMaximas();
+    Integer getCuotasMaximas();
     
     BigDecimal getCapitalMinimo();
